Add ImageProperties...
correct some other classifications.
Add a few type annotations
